requirements.txt
# Galaxy Sommelier Requirements
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Astronomy libraries
astropy>=5.3.0
astroquery>=0.4.6
photutils>=1.8.0
astropy-healpix>=0.7
reproject>=0.10.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Data handling
h5py>=3.9.0
zarr>=2.16.0
tqdm>=4.65.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment tracking
wandb>=0.15.0

# Web framework (for API)
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
PyYAML>=6.0
click>=8.1.0
schedule>=1.2.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
jupyter>=1.0.0 